requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
Django==5.0.6
faiss-cpu==1.8.0
filelock==3.14.0
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.5.0
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
openai==0.28.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==2.7.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
sympy==1.12
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.4
transformers==4.41.1
twilio==9.1.0
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.3
yarl==1.9.4
